libs/view/dashboard/src/lib/+state/dashboard.selectors.ts
import { createFeatureSelector, createSelector } from '@ngrx/store';
import {
  DASHBOARD_FEATURE_KEY,
  State,
  DashboardPartialState
} from './dashboard.reducer';

// Lookup the 'Dashboard' feature state managed by NgRx
export const getDashboardState = createFeatureSelector<
  DashboardPartialState,
  State
>(DASHBOARD_FEATURE_KEY);

export const getDashboardLoaded = createSelector(
  getDashboardState,
  (state: State) => state.loaded
);

export const getDashboardError = createSelector(
  getDashboardState,
  (state: State) => state.error
);
